pyproject.toml
[tool.poetry]
name = "data-service"
version = "0.1.0"
description = "Backend service for serving datafiles over http"
authors = ["Team Microdata"]
license = "Apache"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.2"
pyarrow = "3.0.0"
uvicorn = "^0.15.0"
google-cloud-storage = "^1.42.3"
pandas = "^1.3.3"
json-logging = "^1.5.0-rc0"
numpy = "1.22.0"
python-dotenv = "^0.19.0"
PyJWT = "^2.4.0"
aiofiles = "^0.7.0"
cryptography = "^35.0.0"
tomlkit = "^0.10.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-dotenv = "^0.5.2"
cryptography = "^35.0.0"
pycodestyle = "^2.8.0"
pylint = "^2.14.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
